Se agregan funcionalidades del segundo sprint respecto a agregar alternativas de problema y se avanza en la parte de calificacion y puntuacion

+++ src/main/java/ufps/ahp/model/PuntuacionAlternativa.java
@@ -43,7 +43,7 @@ public class PuntuacionAlternativa implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "valor")
-    private int valor;
+    private float valor;
     @JoinColumn(name = "puntuacion_alternativa_criterio", referencedColumnName = "id_puntuacion_alt_crit")
     @ManyToOne(optional = false)
     private PuntuacionAlternativaCriterio puntuacionAlternativaCriterio;
@@ -71,7 +71,7 @@ public class PuntuacionAlternativa implements Serializable {
         this.idPuntuacionAlternativa = idPuntuacionAlternativa;
     }
 
-    public int getValor() {
+    public float getValor() {
         return valor;
     }
